aisecure/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
python-docx
pypdf
groq
python-dotenv
